Revert "Update conversion functions too"
This reverts commit aa2fc258e6f228175bc8d808e2d762810fda3ab4.

+++ src/modules/proto_dhcpv4/ncc_util.c
@@ -503,19 +503,19 @@ int ncc_host_addr_resolve(ncc_endpoint_t *host_ep, char const *host_arg)
 /*
  *	Convert a struct timeval to float.
  */
-double ncc_timeval_to_float(struct timeval *in)
+float ncc_timeval_to_float(struct timeval *in)
 {
-	double value = (in->tv_sec + (double)in->tv_usec / USEC);
+	float value = (in->tv_sec + (float)in->tv_usec / USEC);
 	return value;
 }
 
 /*
  *	Convert a float to struct timeval.
  */
-int ncc_float_to_timeval(struct timeval *tv, double in)
+int ncc_float_to_timeval(struct timeval *tv, float in)
 {
 	/* Boundary check. */
-	if (in >= (double)LONG_MAX) {
+	if (in >= (float)LONG_MAX) {
 		ERROR("Cannot convert to timeval: float value %.0f exceeds LONG_MAX (%ld)", in, LONG_MAX);
 		return -1;
 	}
